Add Fortran JSON round-trip check (#2652)
Switch the Fortran language module to emit real(real64) literals (and
update the freal stub signature in the fval_m preamble) so double-
precision values like 1.7976931348623157e+308 are no longer silently
truncated to single precision, regenerate the Fortran goldens, add a
real fval_m.f90 with a hand-rolled to_json serializer plus the
run_fortran_roundtrip.py driver and the lint.yml step, and expand the
shared roundtrip input with an empty-string case.

+++ tests/integration/cases/float_scientific_notation/Fortran@v2008.f90
@@ -1,5 +1,5 @@
 module fval_m
-  use, intrinsic :: iso_fortran_env, only: int64
+  use, intrinsic :: iso_fortran_env, only: int64, real64
   implicit none
   type :: fval_t
     integer :: t = 0
@@ -8,7 +8,7 @@ contains
   function fnull() result(v); type(fval_t) :: v; end function
   function fbool(b) result(v); logical, intent(in) :: b; type(fval_t) :: v; end function
   function fint(n) result(v); integer(kind=int64), intent(in) :: n; type(fval_t) :: v; end function
-  function freal(x) result(v); real, intent(in) :: x; type(fval_t) :: v; end function
+  function freal(x) result(v); real(kind=real64), intent(in) :: x; type(fval_t) :: v; end function
   function fstr(s) result(v); character(len=*), intent(in) :: s; type(fval_t) :: v; end function
   function flist(a) result(v); type(fval_t), intent(in) :: a(:); type(fval_t) :: v; end function
   function fmap(a) result(v); type(fval_t), intent(in) :: a(:); type(fval_t) :: v; end function
@@ -20,9 +20,9 @@ program main
     implicit none
     type(fval_t) :: my_data
     my_data = flist([fval_t :: &
-        freal(0.0), &
-        freal(1.0), &
-        freal(1500.0), &
-        freal(0.001) &
+        freal(0.0_real64), &
+        freal(1.0_real64), &
+        freal(1500.0_real64), &
+        freal(0.001_real64) &
     ])
 end program main
