+++ tests/integration/cases/scalars/Fortran_combined@v2008.f90
@@ -1,5 +1,5 @@
 module fval_m
-  use, intrinsic :: iso_fortran_env, only: int64
+  use, intrinsic :: iso_fortran_env, only: int64, real64
   implicit none
   type :: fval_t
     integer :: t = 0
@@ -8,7 +8,7 @@ contains
   function fnull() result(v); type(fval_t) :: v; end function
   function fbool(b) result(v); logical, intent(in) :: b; type(fval_t) :: v; end function
   function fint(n) result(v); integer(kind=int64), intent(in) :: n; type(fval_t) :: v; end function
-  function freal(x) result(v); real, intent(in) :: x; type(fval_t) :: v; end function
+  function freal(x) result(v); real(kind=real64), intent(in) :: x; type(fval_t) :: v; end function
   function fstr(s) result(v); character(len=*), intent(in) :: s; type(fval_t) :: v; end function
   function flist(a) result(v); type(fval_t), intent(in) :: a(:); type(fval_t) :: v; end function
   function fmap(a) result(v); type(fval_t), intent(in) :: a(:); type(fval_t) :: v; end function
@@ -21,7 +21,7 @@ subroutine main_declaration()
     type(fval_t) :: my_data
     my_data = flist([fval_t :: &
         fint(42_int64), &
-        freal(3.14), &
+        freal(3.14_real64), &
         fbool(.true.), &
         fbool(.false.), &
         fstr('hello "world"') &
@@ -34,7 +34,7 @@ subroutine main_assignment()
     type(fval_t) :: my_data
     my_data = flist([fval_t :: &
         fint(42_int64), &
-        freal(3.14), &
+        freal(3.14_real64), &
         fbool(.true.), &
         fbool(.false.), &
         fstr('hello "world"') &
